@@ -25,27 +25,27 @@
   "author": "zcong1993",
   "license": "MIT",
   "devDependencies": {
-    "@grpc/grpc-js": "^1.3.7",
+    "@grpc/grpc-js": "^1.4.1",
     "@types/jest": "^27.0.2",
-    "@types/node": "^16.9.6",
-    "@typescript-eslint/eslint-plugin": "^4.31.2",
-    "@typescript-eslint/parser": "^4.31.2",
+    "@types/node": "^16.11.1",
+    "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.0.0",
     "chalk": "^4.1.2",
     "conventional-changelog-cli": "^2.1.1",
-    "eslint": "^7.32.0",
+    "eslint": "^8.0.1",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.2",
+    "eslint-plugin-import": "^2.25.2",
     "execa": "^5.1.1",
-    "google-protobuf": "^3.18.0",
+    "google-protobuf": "^3.18.1",
     "husky": "^7.0.2",
-    "jest": "^27.2.1",
-    "lint-staged": "^11.1.2",
+    "jest": "^27.3.0",
+    "lint-staged": "^11.2.3",
     "minimist": "^1.2.5",
     "prettier": "^2.4.1",
-    "prompts": "^2.4.1",
+    "prompts": "^2.4.2",
     "semver": "^7.3.5",
-    "ts-jest": "^27.0.5",
-    "typescript": "^4.4.3"
+    "ts-jest": "^27.0.7",
+    "typescript": "^4.4.4"
   },
   "jest": {
     "preset": "ts-jest",
@@ -60,7 +60,7 @@
     ]
   },
   "dependencies": {
-    "rxjs": "^7.3.0"
+    "rxjs": "^7.4.0"
   },
   "peerDependencies": {
     "@grpc/grpc-js": "^1.2.10"
